fix a shit ton of bug

+++ database.py
@@ -207,22 +207,23 @@ def update_collections(new_collections):
 
             # Weed out the shit we gonna replace
             existing_collections={}
+            new_collection_names=[f"★ Collection #{data.get_sources().collection_index[x]}" for x in new_collections.keys()]
             for collection in current_collections["collections"]:
-                if collection["name"] not in new_collections.keys():
+                if collection["name"] not in new_collection_names:
                     existing_collections[collection["name"]]=collection["hashes"]
 
             b.write_uint(len(new_collections)+len(existing_collections))
 
-            # Write the new collections
+            # Write the existing collections
             for name, checksums in existing_collections.items():
                 b.write_string(name)
                 b.write_uint(len(checksums))
                 for checksum in checksums:
                     b.write_string(checksum)
                     
-            # Write the existing collections
-            for name, checksums in new_collections.items():
-                b.write_string(name)
+            # Write the new collections
+            for source_key, checksums in new_collections.items():
+                b.write_string(f"★ Collection #{data.get_sources().collection_index[source_key]}")
                 b.write_uint(len(checksums))
                 for checksum in checksums:
                     b.write_string(checksum)
@@ -230,7 +231,7 @@ def update_collections(new_collections):
             osu_folder=data.get_settings().osu_install_folder
             collection_file=os.path.join(osu_folder,"collection.db")
             backup_file=os.path.join(osu_folder, "collection_backup.db")
-            if os.path.isfile(collection_file):
+            if not os.path.isfile(backup_file) and os.path.isfile(collection_file):
                 shutil.copy2(collection_file, backup_file)
             with open(collection_file, "wb") as db:
                 db.write(b.data)
